Dynamic window generation
Allow for an infinite number of (editor, presenter) window pairings,
using an index based system that is stored by the editor and used with
the presenting user action. If the associated presenter window does not
yet exist, it is created, and then focussed and the content is
transferred for rendering.